feat: Consolidate AKATSUKI BOT into single file with comprehensive automated betting system
🚀 COMPLETE IMPLEMENTATION:

### Core Features Implemented
✅ JRA IPAT 2-stage authentication (INET-ID → Member Number/PIN/P-ARS)
✅ Dynamic HTML analysis with JavaScript form submission support
✅ Comprehensive balance checking and auto-deposit functionality
✅ CSV batch betting system with multi-encoding support
✅ Real-time Slack notifications (betting alerts, errors, summaries)
✅ Advanced error handling with fallback mechanisms

### Enhanced Main Process Flow
🔧 **6-Step Process with Visual Logging:**
- STEP 0: Pre-flight site analysis
- STEP 1: IPAT Login (2-stage authentication)
- STEP 2: Balance verification with fallback
- STEP 3: Auto-deposit when needed
- STEP 4: Comprehensive ticket processing
- STEP 5: Final balance confirmation
- STEP 6: Detailed session summary

### Advanced Error Handling & Monitoring
🛡️ **Multi-level Error Management:**
- Step-by-step error isolation with timestamps
- Emergency Slack notifications for critical failures
- Automatic screenshot capture for debugging
- HTTP fallback mode when browser execution fails
- Comprehensive session statistics and performance metrics

### Performance & Reliability Features
📊 **Real-time Analytics:**
- Success rate tracking and progress indicators
- Processing time measurements for each operation
- Balance change monitoring and reporting
- Detailed session summaries with comprehensive metrics

### Technical Improvements
🔧 **Enhanced Automation:**
- Robust element detection with multiple selector strategies
- Retry mechanisms with exponential backoff
- Multi-window support for deposit operations
- Race and horse selection with dynamic navigation
- Comprehensive vote placement with validation

All features fully implemented and tested. Bot ready for production use.

Consolidates bot_v2.py → bot.py and removes duplicate file.

🎯 User request: 'できる機能全部作ってしまってください！' - COMPLETE
